Fanning Weapon and Prices
Fanning Weapon is now a tactic instead of a technique. This renders it unique to fighters instead of general. A question can be raised whether this is a good thing, but note that tricksters and fighters now have features directly doing this effect, and the rest of the classes don't exactly need further help for it (be it with sentinels not seeking this playstyle and transmuters have more than enough alchemy support).
In addition, Refreshening Straw was adjusted to treat more conditions, and costs less.